+++ packages/server/src/modules/shop/shop.module.ts
@@ -2,39 +2,13 @@ import { Module } from '@nestjs/common';
 import { ShopService } from './shop.service';
 import { ShopController } from './shop.controller';
 import { MongooseModule } from '@nestjs/mongoose';
-import { Shop, ShopSchema } from './schemas/shop.schema';
-import { MulterModule } from '@nestjs/platform-express';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import * as multerS3 from 'multer-s3';
-import { MulterOptions } from '@nestjs/platform-express/multer/interfaces/multer-options.interface';
-import { S3Client } from '@aws-sdk/client-s3';
-
-const multerOptionsFactory = (configService: ConfigService): MulterOptions => ({
-  storage: multerS3({
-    s3: new S3Client({
-      region: configService.get('AWS_S3_REGION'),
-      credentials: {
-        accessKeyId: configService.get('AWS_S3_ACCESS_KEY_ID'),
-        secretAccessKey: configService.get('AWS_S3_SECRET_ACCESS_KEY'),
-      },
-    }),
-    bucket: configService.get('AWS_S3_BUCKET'),
-    contentType: multerS3.AUTO_CONTENT_TYPE,
-    cacheControl: 'max-age=31536000',
-    key: function (req, file, cb) {
-      cb(null, Date.now().toString());
-    },
-  }),
-});
+import { Shop, ShopSchema } from '../../schemas/shop.schema';
+import S3MulterModule from '../../providers/S3MulterModule';
 
 @Module({
   imports: [
     MongooseModule.forFeature([{ name: Shop.name, schema: ShopSchema }]),
-    MulterModule.registerAsync({
-      imports: [ConfigModule],
-      useFactory: multerOptionsFactory,
-      inject: [ConfigService],
-    }),
+    S3MulterModule(),
   ],
   controllers: [ShopController],
   providers: [ShopService],
